requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
nltk==3.8.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
kaggle==1.5.16
wordcloud==1.9.2
plotly==5.15.0
sentence-transformers==2.2.2
networkx==3.1
python-dotenv==1.0.0